test(mcp-client): run MCP client workflow tests against both stdio and http
Extracts shared MCP tool-call assertions (create-collection, index,
search, facets, pagination, schema) into McpClientIntegrationTestBase.

McpClientIntegrationTest (http) extends the base and boots the full
Spring application with @SpringBootTest + HttpClientStreamableHttpTransport.

McpClientStdioIntegrationTest (stdio) extends the base without a Spring
context: spawns the assembled jar as a subprocess via StdioClientTransport
and connects a Testcontainers-managed SolrContainer through the env var.
Both transports exercise the same 15-test workflow, verified on JDK and
GraalVM nativeTest.
